requirements.txt
streamlit
chromadb
sentence-transformers
pypdf
python-dotenv
openai